403実装しました

+++ backend/app/Http/Controllers/Api/DeleteScheduleController.php
@@ -14,7 +14,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Http\Exceptions\HttpResponseException;
 use App\Http\Controllers\Api\Content;
-
+use Illuminate\Support\Facades\Auth;
 
 class DeleteScheduleController extends ApiController
 {
@@ -39,6 +39,17 @@ class DeleteScheduleController extends ApiController
         //db内にこのレコードのデータ歩かないかをフォームリクエストにある
         //softdalete Laravel標準の論理削除
         try{
+
+            $user_id = Auth::user()->id;
+            $result = Schedule::where('user_id',$user_id)->where('id',$id)->get()->first();
+            if(is_null($result)){
+                return response()->json([
+                    'success' => false,
+                    'code' => 403,
+                    'message' => "認可されてないよ"
+                ], 403);
+            }
+
             $schedule = Schedule::find($id)->delete();
             return response()->json([
                 'success' => true,
